Django-11/projects/QuizApp/requirements.txt
asgiref==3.5.2
autopep8==1.7.0
dj-rest-auth==2.2.5
Django==4.1.1
django-nested-admin==3.4.0
djangorestframework==3.14.0
pycodestyle==2.9.1
python-decouple==3.6
python-monkey-business==1.0.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
